backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# 任务队列
celery==5.3.4
redis==5.0.1

# 数据库
sqlalchemy==2.0.25

# HTTP客户端
httpx==0.26.0
requests==2.31.0

# 视频处理
moviepy==1.0.3
opencv-python==4.9.0.80
pillow==10.2.0

# 音频处理
pydub==0.25.1
librosa==0.10.1

# AI模型
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
transformers==4.37.0
faster-whisper==0.10.0

# 字幕处理
stable-ts==2.15.0

# 工具
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
